requirements-heroku.txt
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.24
requests>=2.31
python-dotenv>=1.0
celery>=5.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
google-api-python-client>=2.0
requests-oauthlib>=1.3.0

# SMS integration
twilio==8.10.0
clicksend-client>=5.0.0

# AI integration
openai>=1.0
elevenlabs>=2.9.0

# Audio processing
mutagen>=1.47.0

# Web server
Flask==2.3.3
fastapi>=0.100
uvicorn>=0.20
gunicorn>=21.0
websockets>=11.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

# Web scraping
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database storage (no additional dependencies needed)
